update gql package with some tweaks and comments
