--- pkg/gql/enum.go
@@ -1,64 +1,0 @@
-package gql
-
-type Enum struct {
-	Name        string
-	Description string
-	Directives  Directives
-	Values      EnumValues
-}
-
-func (e *Enum) GetDescription() string {
-	return e.Description
-}
-
-func (e *Enum) GetName() string {
-	return e.Name
-}
-
-func (e *Enum) GetDirectives() []Directive {
-	return e.Directives
-}
-
-func (e *Enum) GetKind() TypeKind {
-	return EnumKind
-}
-
-func (e *Enum) GetValues() []*EnumValue {
-	return e.Values
-}
-
-// TODO: enum values has to be a list, return
-
-// var _ EnumType = &Enum{}
-
-type EnumValues []*EnumValue
-
-type EnumValue struct {
-	Name        string
-	Description string
-	Directives  Directives
-	Value       interface{}
-}
-
-func (e EnumValue) GetDescription() string {
-	return e.Description
-}
-
-func (e EnumValue) GetDirectives() []Directive {
-	return e.Directives
-}
-
-func (e EnumValue) GetValue() interface{} {
-	return e.Value
-}
-
-func (e EnumValue) IsDeprecated() bool {
-	for _, d := range e.Directives {
-		if _, ok := d.(*deprecated); ok {
-			return true
-		}
-	}
-	return false
-}
-
-// var _ EnumValue = &EnumValue{}
