+++ pkg/gql/errors.go
@@ -1,7 +1,13 @@
 package gql
 
+/*
+Errors is an alias for a bunch of Error
+*/
 type Errors []*Error
 
+/*
+Error for the Result of the validation/execution
+*/
 type Error struct {
 	Message    string                 `json:"message"`
 	Locations  []*ErrorLocation       `json:"locations"`
@@ -9,15 +15,24 @@ type Error struct {
 	Extensions map[string]interface{} `json:"extensions,omitempty"`
 }
 
+/*
+Error implements the error interface
+*/
 func (e *Error) Error() string {
 	return e.Message
 }
 
+/*
+ErrorLocation represents the location of an error in the query
+*/
 type ErrorLocation struct {
 	Line   int `json:"line"`
 	Column int `json:"column"`
 }
 
+/*
+NewError ...
+*/
 func NewError(msg string, extensions map[string]interface{}) CustomError {
 	return &cErr{
 		msg:  msg,
@@ -49,10 +64,10 @@ type CustomError interface {
 }
 
 const (
-	ErrValidateOperationName                  = "Operation name '%s' is defined multiple times"
-	ErrAnonymousOperationDefinitions          = "Can not use anonymous operation where multiple operation definitions exist"
-	ErrLeafFieldSelectionsSelectionNotAllowed = "Selection on type '%s' is not allowed"
-	ErrLeafFieldSelectionsSelectionMissing    = "Selection on type '%s' is missing"
-	ErrFieldDoesNotExist                      = "Field '%s' does not exist on type '%s'"
-	ErrResponseShapeMismatch                  = "fields in set can not be merged: %s"
+	errValidateOperationName                  = "Operation name '%s' is defined multiple times"
+	errAnonymousOperationDefinitions          = "Can not use anonymous operation where multiple operation definitions exist"
+	errLeafFieldSelectionsSelectionNotAllowed = "Selection on type '%s' is not allowed"
+	errLeafFieldSelectionsSelectionMissing    = "Selection on type '%s' is missing"
+	errFieldDoesNotExist                      = "Field '%s' does not exist on type '%s'"
+	errResponseShapeMismatch                  = "fields in set can not be merged: %s"
 )
